Bug#31354435: DELETE VESTIGIAL CODE IN FIL_SHARD::GET_FILE_FOR_IO()
There is a section of code in Fil_shard::get_file_for_io() that was only
needed when undo truncation used the same space ID before and after the
truncate.

This code had a comment that said "Page access request for a page that is
outside the truncated UNDO tablespace bounds", indicating that the undo
tablespace previously had the page but now it does not.  That cannot
occur now that the old space is replaced by a new space with a different
space ID.

In order to prove that this code does not occur anymore, I added an
ib::fatal() message and ran it in PB2 mysql-8.0-stage without it hitting.

I also looked at the testcase called innodb_undo.undo_ddl_vs_dml to try
to watch what happens when a lot of DML occurs while undo tablespaces are
being truncated. This testcase was not doing what it should, so I improved
it.  Now, 4 concurrent threads make 4 undo tablespaces active and then
inactive while doing a bunch of updates. The threads offset their respective
activity so that the timing of undo truncation does not coincide. And each
thread acts upon different records within the table.

Approved by Kuba in RB#24477