Bug#30435378 CLEAR DD CONTENTS DURING SYNC AFTER INITIAL CLUSTER RESTART
Problem:
--------

An initial Cluster restart deletes all NDB metadata (and data).
A subsequent MySQL Server (re)connection sees the removal of all
NDB tables from the DD. Logfile groups and NDB tablespaces, on
the other hand, are removed as part of the larger synchronization
operation which compares the contents of NDB Dictionary and DD on
a per-object basis.

Fix:
----

Removes all logfile groups and NDB tablespaces in addition to
NDB tables from the MySQL Server DD during synchronization when
an initial restart is detected to have taken place.

Note that the old behaviour of removing DD table definitions
corresponding to NDB tables that have been removed from NDB
Dictionary is retained for "normal" system restarts. This is
not extended to remove logfile groups and tablespaces.

Change-Id: I7bb00004f5de528101e9aac577fd107b972c4a61