Bug#31328947 WRONG SHARED GLOBAL MEMORY PAGES DEDICATED FOR HIGH PRIORITY ALLOCATIONS.
Description:
-----------
In data nodes page oriented allocations are divided into nine resource groups
which have some pages dedicated for themselves and there are some pages,
shared global memory, that can be allocated by any resource group.

To not have the Query memory resource group starve out other more "important"
resource groups, like transaction memory resource, allocations for query
memory is taken as lower priority and are not allowed to use the last 10% of
shared global memory.

That change was introduced by WL#12523 "Poolification: limit query memory in
Dbspj" in 8.0.15.

Although the calculation of the number of pages of shared global memory that
QM may not use was wrong:

    m_prio_free_limit = (m_allocated - m_free_reserved) * HIGH_PRIO_FREE_PCT
/ 100;

The above formula is correct if there are no pages in use, which happen to be
true if LateAlloc is off.

If LateAlloc is on, the allocation of page memory is done in two steps.

First a smaller amount enough to proceed with node start is allocated, and
later all the rest of page memory is allocated, and when using the above
formula at that point several pages are typically in use and formula is not
valid.

How To Repeat:
-------------
After applying Bug#31231286 Improve debug info on page allocation failure
[noclose], look on test failures in PB2.

Fix:
----
Use get_shared() function to get amount for shared global memory and use that
to calculate the amount of pages reserved for high priority pages.

    m_prio_free_limit = get_shared() * HIGH_PRIO_FREE_PCT / 100 + 1;

Reviewed-by: Sanjana DS <sanjana.ds@oracle.com>