Bug #31217037 MYSQL ERROR LOG GETS FILLED WITH SAME ERROR MESSAGE CAUSED BY COMPONENT DEINIT
Description:
At the server shutdown time if there exits any loaded component
deinit function failure, it calls the same deinit function infinity
times which leads to fill the server error log file, with the
same error message. Due to this the disk is getting filled up.

The error message it prints to the error log is as below:
2020-04-23T07:18:20.045708Z 0 [ERROR] [MY-013129] [Server] A message
intended for a client cannot be sent there as no client-session is
attached. Therefore, we're sending the information to the error-log
instead: MY-003538 - Unknown error 3538

Analysis: mysql_dynamic_loader_imp::deinit is called at mysql server
shutdown time. In this function the loaded components will be
unloaded as a component groups. In the process of unloading the
component deinit function will be called first and then unload
of the component happens.
The bug is seen when the deinit function returns 'true', then it has
to print the error message and has to continue with the rest of the
components in the list.
But in the mysql_dynamic_loader_imp::deinit() function the
unload_do_topological_order() return value is not captured, and it
is allowed to call repeatedly for the same component, which leads to
print the error message infinity times to the error log.

Fix: When unload_do_topological_order() call recieves error,
we need to remove that component group node from the farword_list
and go further. In the error message the component name is not
printed, that is becuase in clean_up() the error unregister is
happening before the component_infrastructure_deinit().
So to get the component names, the error unregister code is moved to
after component_infrastructure_deinit().

RB: 24318
Reviewed by : Joro <georgi.kodinov@oracle.com>
              Marek <marek.szymczak@oracle.com>