Bug #31326120 UBSAN: SHIFT EXPONENT 4294967294 IS TOO LARGE FOR 64-BIT TYPE
The basic lookup function find_type() in typelib returns an int -1 in
case of ambiguities.  When this is converted to unsigned, and used as
an argument for left-shift, we get undefined behaviour.

Propagate the -1, and handle it as an error before doing any left-shifting.

Change-Id: Ia8508889a0dfb5fc044f37f65ce1544697aa486e
(cherry picked from commit a621c19d7e1a5327982d08a1550968989754cea9)