Bug#30899683 INNODB CRASH WHEN DROP TABLE IN DEBUG MODE
Problem
--------
MySQL server crashed while dropping a table created with an
invalid COMPRESSION clause such as "COMPRESSION='zlibX'" with
INNODB_STRICT_MODE=OFF

The crash was because the InnoDB did not return error on encountering
a table created with invalid COMPRESSION option.

Solution
--------
Return error to the caller to be handled gracefully.

RB: 24434

Reviewed By: Sunny Bains <sunny.bains@oracle.com>
             Sivert Sorumgaard <sivert.sorumgaard@oracle.com>