Bug#31347684 BREAK NDB_RESTORE DEPENDENCY ON NDBCLUSTER PLUGIN
The ndb_restore utility depend on files which are part of the ndbcluster
plugin. This causes a risk that changes to ndbcluster plugin
inadvertently affects the functionality of ndb_restore.

The ndb_dist_priv_util.h is included in order to keep track of which
table names it should not restore.

Remove dependency by listing table names to exclude directly in
restore_main.cpp

Change-Id: Ic4d0cd5b6c01bd737d39c356199a1dd1b1d3ab0f