Bug#29899151: CORE AT CREATE TABLE WITH FK FOR TABLE TRANSFERRED FROM MYISAM TO INNODB
Problem: If the creation of a child table in a FK-relation triggered
an engine substitution, an assert would result. Root cause was that
engine substitution was done after decisions had been taken based on
values form the handlerton object representing the original
engine. This then lead to internal inconsistencies.

Solution: Move the checking of enabled/disabled engines from
check_engine() to to a separate function which is called earlier,
before the type of engine has had an effect. Moving the call to
check_engine() itself was not possible since some of the other checks
could not be moved.

In addition, refactor the code which tracks disabled storage engines
so that it no longer requires boost.

Change-Id: Id13de85f4180568c7d7b205315687aa192feb591