Bug #30589827 	MYSQL-8.0: MASTER NODE CRASHED DURING SHUTDOWN WITH AN ERROR "CHECK PTR.P->MASTE
The master node in a backup crashes on receiving duplicate replies to
a DEFINE_BACKUP_REQ. The duplicate replies occur because a non-master
node fails during the backup. The backup master detects the node-failure
and handles it by sending itself a DEFINE_BACKUP_REF on behalf of the
failed node. However, the failed node has already sent a DEFINE_BACKUP_CONF
to the master before crashing, so two replies are received from the same
node - one CONF from the failed node and one REF from the master on behalf
of the failed node. The backup master attempts to enforce an ndbrequire
that only one reply per node is expected and crashes on failure. This
scenario of duplicate replies and crashes can also occur for
START_BACKUP_REQ and STOP_BACKUP_REQ.

This is fixed by allowing duplicate replies if the error is a node-failure.
Test cases are added for DEFINE_BACKUP_REQ, START_BACKUP_REQ and
STOP_BACKUP_REQ.

Change-Id: I987199f22ee0352510fb670853cd55b0240e8c16