Bug#30819012 - MYSQLPUMP VALIDATES ALL DATABASE OBJECTS BY                IGNORING --EXCLUDE_DATABASES AND                --SINGLE-TRANSACTION OPTION
DESCRIPTION
===========
While trying to dump a single schema using mysqlpump, it
validates all database objects irrespective of the specific
database instructed to export using --exclude-databases
and --single-transaction.

ANALYSIS
========
According to the current mysqlpump design, it first
enumerates all the DB objects as part of the
Crawler module and later performs the filtering of the
objects before dumping them. This causes all the database
objects to be validated even if a single database was
instructed to be dumped.

FIX
===
The fix performs object filtering before enumeration
so that only the databases and their corresponding
objects that should be included in the dump are validated.

RB# 24340