Bug#31213556 AUTOTEST TESTFK -N DROPTABLEWITHFKDURINGRESTART FAILS SPORADICALLY
The testcase verifies that any foreign key created or dropped during a
data node's downtime is properly handled by the data node when it comes
back online. When the testcase starts the data node to verify this, it
is not waiting for a sufficient amount of time for the node to start.
Due to this the test fails sporadically in autotest. This patch fixes
this by increasing the timeout value used by the testcase to wait for the
node start.

Change-Id: Ica2bdfd5058794d1855b890ad60695a8074371c6