Bug #31355892 AUTOTEST : TESTSYSTEMRESTART -N BUG45154 D1 TABLEID WAS NOT REUSED
Test requires that 2 different tables are created with the same
tableid to get coverage of the fix it tests.

The second table has a different distribution, so requires the
creation of a new Hashmap when it is created, if it does not
already exist.

So the first time the test is run, the 2nd table will generally
have a different tableid, as the new Hashmap has taken the lower
id.  Subsequent runs will pass.

Solution is to create both tables once up-front, so that all
needed hashmaps are created up-front.

Approved-by : JD Duncan <john.duncan@oracle.com>