Bug#31364756 JSON NDB TESTS SKIPPED ON NDBCLUSTER RUNS
Problem:
--------

The ndbcluster runs on mysql-8.0 and mysql-trunk Server push build branches
run all the default MySQL Server tests with the MySQL Cluster build enabled.
In addition, a small subset of NDB tests are run including tests that are
especially prone to regressions due to changes in the MySQL server code.

The json tests, which are designed to use a single include file in multiple
SE specific tests, were skipped even though they are prone to regressions.

Fix:
----

Tweak the collections file to run the json NDB tests as well

Change-Id: Ib4663ede2f648e3a324a02e11de02e2ee22c4918