Bug#31370227 BUG IN THE DOUBLEWRITE CODE IN FUNCTION  DOUBLE_WRITE::SYNC_PAGE_FLUSH
Issue
=====
If the write to datafile has failed as part of doublewrite's flush and sync
operation we want the system to continue working normally. But this doesn't
seem to the case currently as we release back the file segment required for
single page writes only if the write to datafile has succeeded.

Fix
===
Fix the issue by allowing the segment release regardless of the status of write
to datafile.

RB: Reviewed on slack
Reviewed-by: Sunny Bains <sunny.bains@oracle.com>