Bug#31017863 NO REFRESH OF METADATA WHEN PRIMARY CHANGES AND GR NOTIFICATIONS ARE USED
The Router is wrongly assuming that each new GR change that is notified by
the x-protocol notifications has new view_id. That is not a case for changes
like switch of the primary or change of the role. The Router can't use the
view_id for the notification debouncing.

This patch fixes this issue. The view_id is no longer taken into account when
the Router gets the GR notification from the x-plugin.

RB: 24473