Bug #31360025 UPDATE TO LIBEDIT 20190324-3.1 BREAKS CONTROL-C BEHAVIOUR IN MYSQL CLIENT
The patch for
    WL#13534 Update bundled libedit to 20190324-3.1
broke backwards compatibility for handling control-C in the bundled
version of libedit.

In order to restore earlier behaviour: control-C will abort the
current input line, and output a new prompt, we need to patch readline
to handle SIGINT, and return to the application.

Note that this means that "bundled" and "system" libedit will have
(slighly) different behaviour.

Change-Id: I76a1aefa850699056d99402dca1eb5c27cf57dca