Bug#30765952 AUTO INCREMENT VALUE DIDN'T RESET CORRECTLY
Issue
=====
During copying of two tables in alter table, although
ha_innobase::get_auto_increment() can use InnoDB's estimation of rows stats in
the table for calculating the auto-increment values, it along with server can
still work out the correct auto increment values even if the estimation is
incorrect. The logic assumes that the very first time the server expects an
autoinc value for a column in a table, the trx->n_autoinc_rows (which stores
the info about number of autoinc rows required for the multi-row inserts) is
required to be zero. Once this value gets updated, after each bulk insert this
counter is decremented and in the case of correct estimation the counter would
be 0 by the end of the bulk insert. But in the case that the estimation is
incorrect it's possible that trx->n_autoinc_rows may not be 0 when bulk insert
has ended, and this needs to be cleared so that the next multi-row insert can
start afresh. But this cleanup doesn't seem to be happening currently
resulting in the reported failure.

Fix
===
Do the required cleanup after each bulk insert.

RB: 24492
Reviewed-by: Debarun Bannerjee <debarun.banerjee@oracle.com>