Bug #31326231: PROTOCOL_COMPRESSION_ALGORITHMS CAN BE SET TO EMPTY, AND SERVER ONLY                ACCEPTS ZLIB CONNECTION
Problem: protocol_compression_algorithms variable specifies what protocol
compression algorithms server supports, if this variable is set to empty
string then server should not allow any connection, instead it allows
connection.

Fix: Setting protocol_compression_algorithms to empty string is not valid.
If server does not want compression to be supported then there is a value
"uncompressed" for it. Fixed it by reporting error when this variable is
set to empty string.

RB#24439