Bug#31408095 WRONG INDEX ENTRY RELEASED WHEN DROPPING INDEX
The wrong index is used in the array of indexes while dropping an index.
For a table with 64 indexes this causes uninitialized memory to be
released. Crash occurs in debug mode when dropping index on a table with
64 indexes. Release compiled may have crashed with double free or
similar. Since the wrong index entry is actually always released, this
problem would also have caused a memory leak when a new index was later
created .

Fix by using the proper array index when releasing the dropped index
entry.

Change-Id: I40fcb34170b6a2fbbd133c0d397046059d28e6bb