Bug#31414848: REMOVE PREPARE_LOAD_TABLE FROM HANDLER API AND MOVE FUNCTIONALITY TO LOAD_TABLE
Problem: Tables which had been added to CP could be left forever if the
load failed. The zombie table could create problems if a table with the
same name, but different definition was added later.

Solution: Remove prepare_load_table which is no longer needed, and add
tables to CP in load_table. Do this after the check for an already
loaded table has beens done, and use a scope guard to ensure that the
added table is removed in case of errors later in load_table.

Change-Id: I36965cabf3e6928cfa8d2ef7a03c1fa28c5b807d