Bug#31242689 MYSQL-8.0 (8.0.21): MASTER CRASHED WITH ERROR "HA_NDBCLUSTER_BINLOG.CC:334: VOID
In Unix based operating systems, Binlogs can be flushed by sending SIGHUP
signal to the server. when the logs are flushed by sending SIGHUP, as sql
command is empty, it will point to a default value(SQLCOM_END). But,
Ndbcluster binlog only expects Flush, Reset and Show binlog commands to
pass through. So, this leads to an assertion failure in ndbcluster.

As it is a valid scenario and keeping in mind that it is possible to get
similar cases like this in the future, fixed the issue by removing the debug assert.

Also added test cases to check the same.

Change-Id: I09fb9ceace210ecf377625e7bff2a76b1b2aebbf