Bug#31422721 MYSQLROUTER BOOSTRAP FAILS WITH COLLATION ISSUE
In Metadata schema versions higher than 1.0.1
instances.mysql_server_uuid uses the collation ascii_general_ci which
then when doing comparisons with the sysvar @@server_uuid results
in an illegal mix of collations.

This patch does the right cast to avoid the collation mismatch in the queries
where server_uuid variable is compared to the metadata field.

RB: 24558