Bug #30187949 SEND GCP_SAVECONF TO BLOCK 0
From 7.6 onwards, each LDM maintains a local LCP control file.
During restart, this control file is updated in response to
each GCP_SAVEREQ, even when the node is still sending GCP_SAVEREF
in response as it is not yet durable.

On reaching start phase 9, the node needs to start participating
in GCP_SAVEREQ/CONF.  This requires a final write to the local LCP
control file, and then sending a GCP_SAVECONF.

The following sequence of events is problematic :
 1 GCP_SAVEREQ arrives, prior to startphase 9, results in
   GCI write to LCP control file starting
 2 Startphase 9 is entered
 3 LCP control file GCI write completes
 4 LCP control file Restart complete write starts
 5 LCP control file Restart complete write completes
 6 Normal GCP_SAVEREQ processing invoked
   - All log parts flushed + fsynced
 7 Attempt to send GCP_SAVECONF is made

At step 1, the sending DIH info was not saved (as this code path was
originally always immediately sending a GCP_SAVEREF), which means it
is not available if necessary at step 7, resulting in this crash.

The solution is to initialise the GCP/DIH information at step 4, when
it is apparent that a GCP_SAVECONF will result.

Approved-by : Mikael Ronstrom <mikael.ronstrom@oracle.com>